Rename all files to kebab-case. Fix all references.
