+++ packages/backend/src/db/transaction-provider.ts
@@ -2,7 +2,7 @@ import * as Promise from 'bluebird';
 import { Transaction } from 'anydb-sql-2';
 import { BaseService, ServiceContext } from '@h4bff/core';
 import { Database } from './database';
-import { TransactionCleaner } from './transactionCleaner';
+import { TransactionCleaner } from './transaction-cleaner';
 
 /**
  * Provides transaction instance to support transactional behavior within service context.
