+++ packages/backend/src/request/index.ts
@@ -1,2 +1,2 @@
-export * from './requestContextProvider';
-export * from './requestInfo';
+export * from './request-context-provider';
+export * from './request-info';
