+++ packages/backend/src/rpc/index.ts
@@ -1,4 +1,4 @@
-export * from './serviceRegistry';
+export * from './service-registry';
 export * from './dispatcher';
 export * from './middleware';
 export * from './response';
