+++ packages/backend/src/rpc/service-registry.ts
