+++ packages/example/src/nested-apps-plugin.ts
