+++ packages/frontend/src/router/index.ts
@@ -1,3 +1,3 @@
-export * from './routeProvider';
+export * from './route-provider';
 export * from './router';
 export * from './link';
